Reimplemented the Kalman filter
Added the process noise factor to the Kalman filter.  It now works, sort of.  I wouldn't use it to control a UAV but it will suffice for gross altitude measurements.  Also, discovered a bug where I was reading the wrong register for Status on the BME 280, however, the BME 280 status register appears to be a bit funky so I removed the status check and it is now in line with the Bosch device driver which ignores the Status register.  Sigh.